+++ crates/rari/src/rsc/mod.rs
@@ -1,3 +1,4 @@
+pub mod actions;
 pub mod components;
 pub mod flight;
 pub mod rendering;
@@ -17,6 +18,7 @@ pub enum ComponentValue {
     Null,
 }
 
+pub use actions::{handle_form_action, handle_server_action, validate_redirect_url};
 pub use components::{
     ComponentContext, ComponentProp, ComponentRegistry, ComponentType, TransformedComponent,
 };
