+++ crates/rari/src/server/core/mod.rs
@@ -1,7 +1,7 @@
 use crate::error::RariError;
+use crate::rsc::actions::{handle_form_action, handle_server_action};
 use crate::rsc::rendering::core::ResourceLimits;
 use crate::runtime::utils::DistPathResolver;
-use crate::server::actions::{handle_form_action, handle_server_action};
 use crate::server::cache::response;
 use crate::server::config::Config;
 use crate::server::handlers::api::{api_cors_preflight, handle_api_route};
