+++ crates/rari/src/server/mod.rs
@@ -1,4 +1,3 @@
-pub mod actions;
 pub mod cache;
 pub mod compression;
 pub mod config;
